Refactor: TextInput, TextInputType
