+++ app/src/main/java/com/knu/cloud/components/text_input/TextInputType.kt
@@ -1,0 +1,5 @@
+package com.knu.cloud.components.text_input
+
+enum class TextInputType {
+    ID, PASSWORD, FIELD
+}
